Implement CRUD for TodoItem in localStorage
Chose to provide only one utility function in the TodoItem class to
support using localStorage, as the previous two were unnecessary
and seemed clunky to use.

A side effect of this is that we store TodoItem as raw as possible:
with internal properties (those prefixed by _) and the entire
priority & status objects.