Refactor: move view methods to internalControl
Javascript classes provide methods to view data in a semantic way
already via getter methods, so the view methods were deemed superfluous.

This change also delegates the "data packaging" to internalControl; that
is, to have it construct the data in a way accessible to displayControl.

(One may argue then that the utility "serialize" & "deserialize" methods
in Project and "revive" method in TodoItem belong to storageControl, but
these are much more involved than the view methods and have no semantic
alternative, so we leave them as is.)