Adjust how content rendering is done
Remove functions rendering individual components and instead am just
using displayControl.renderDisplay. An advantage of this is simplicity.
A disadvantage of this is that expanded todos do not retain their expanded
state on each call of renderdisplay. This can potentially be solved by
using sessionStorage to store which todos are expanded, so they can be
re-rendered in their expanded form.