Fix project selector disabled state persisting
The bug: clicking to edit a todo causes the project <select> to be
disabled as intended, but the disabled state persisted even after
the form was exited.

The current fix was implemented. An alternative considered was listening
for the "reset" event on the form which was called on form submittal or
pressing the cancel button, but the problem was that pressing the Esc
key to close the form circumvented this listener.